move and rename abstract handler. Make exception mandatory and clean support method
